RFC: add a simplified Push wizard
This is not yet ready (String externalization, tests, etc), but I'd
like to get some feedback: a new "Push to upstream" menu action is
added to the team menu as well as the Repositories View. When clicked,
it tries to push and shows the push result unless it is clear that
push can not work (no upstream configuration and no "origin" remote,
for example), in which case a simple wizard is opened to guide the
user through configuration of the remote configured for the branch (or
"origin" if no such configuration exists).
What is still missing is some smart way of editing the Push
Specification, perhaps some buttons like "Suggest Gerrit pattern" or
such should be available...

Change-Id: I03c9010717117eaa4f6f00ceee20a5037c024e9e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>