[historyView] Abbreviate commit-id in commit list
To gain some screen real estate abbreviate the commit-id in the commit
list. We abbreviate to fixed length of 8 digits in the list to not
waste CPU cycles on doing a more precise abbreviation. The full
commit-id is available in the commit detail pane and can also be
copied to the clipboard.

Change-Id: I96e7c16de62532b4e818b2ba4ef5dde3d3eec86e
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>