Simplify Fetch and Push second part: Add a simplified Push wizard
A new pair of "Push to upstream"/"Fetch from upstream" menu actions is
added to the team menu as well as the Repositories View. When clicked,
the relevant remote is detected (either upstream configuration of the
currently checked out branch, or "origin"), and the push/fetch is
tried for this configuration if possible. If it is clear that
push/fetch is not configured correctly, e.g. there is no URI, a
somewhat simplified wizard is opened to guide the
user through configuration of the remote configured for the branch (or
"origin" if no such configuration exists).
In any case, after the fetch/push, the user can click "Configure..."
on the push/fetch result dialog in order to inspect or change the
configuration.
What is still missing is some smart way of editing the Push/Fetch
RefSpecs, perhaps some buttons like "Suggest Gerrit pattern" or such
should be available...

Bug: 312586
Bug: 313862
Change-Id: I03c9010717117eaa4f6f00ceee20a5037c024e9e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>