Created an extension point to extend the commit message.
To integrate Mylyn's commit templates an extension point
is needed. This extension point provides an interface you
must implement. During the preparation of the commit
dialogs the extensions using this extension points will be
requested for a message to add to the dialog.
Here is a sample configuration:
<extension
    point="org.eclipse.egit.ui.commitMessageProvider">
    <commitMessageProvider
        class="org.eclipse.egit.mylyn.CommitMessageProvider">
    </commitMessageProvider>
</extension>
The mylyn integration is done. You will find them in the
org.eclipse.egit.mylyn project. The implementation of the
ICommitMessagProvider has the ability to check if there are a
project-specific configuration for mylyn's commit message templates.
Change-Id: I3e0b8cd23be957c8c6a8199ce6818282fb24dc68