Handle simple tags in RepositoryUtils
The decorators try to find tags for the branch label, but fail
with an IncorrectObjectTypeException if a tag is a simple
non-annotated tag. For those and annotated tags without a time
stamp, pick the committer time as reference for selecting the
right one.

For simple tags or tags without a timestamp, use the committer
time stamp for sorting.

Bug: 334078
Change-Id: Ide7733d5d7c6a87d15d1ebda20a4f7a88a1a926a
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>