Update patch set 1
Patch Set 1:

One would think that /etc/gitconfig would be the system wide configuration file. I did that for a while, but found out it wasn't, which is also documented in the man page for git-config.

That you cannot save to a non-writeable file is not something I can do anything about, but the user does get an error message. I can annotate the location and insert a read-only warning

Patch-set: 1