Add a "BasicConfigurationDialog"
Note: this needs JGit change

http://egit.eclipse.org/r/#change,2231

to build.

This is a dialog asking the user to confirm/edit their name and 
e-mail once. This allows them to overwrite the system defaults and
store these values in the user-scoped configuration file.
The dialog will only appear once unless the user explicitly deselects
the "Do not show again" check box.
The information that the dialog was already shown is stored as a
Preference in the workspace. It is also possible to open the 
user-scoped Git configuration from this dialog to do more
configuration there.
Since there is no way to invoke this "centrally", this dialog must
be added to all operations that (potentially) create a commit.
The following actions have already been adjusted:

I. Team Actions:
	1. Commit
	2. Merge
II. History View:
	1. Revert
III. Repositories View:
	1. Merge
	2. Rebase
	
Change-Id: Ibe8746aa856b2eae44de63b2986be93fcfccef2d
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>