Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

Yes, we need support for the lightweight tags eventually, but I'm afraid if we only show them in the project label, we'll confuse our users. Before we show the tag in the label, we should first include them in the tag dialog (see bug 323607) and if at all possible in the history view (I'm not aware of a bug for this one...).

Patch-set: 5
Label: SUBM=+1