Check FS.supportsExecute() before resolving file from resource
The performance of building a list of ResourceEntries is improved by
checking the file system's support for executable files before
checking the actual file as this requires resolving the file first -
which seems to be expensive when many resources exist. This only
affects platforms that do not support executable files, e.g. Windows.
See bug 325393 for performance results.

Bug: 325393
Change-Id: I6b353ad1ad184cc0102ee430cc93d85bd29294ea
Signed-off-by: Philipp Thun <philipp.thun@sap.com>