CreateBranch: make overwriting suggested branch name easier
When creating a new branch based on a remote branch, we suggest
the stripped name as local branch name. If the local branch already
exists, the user can not continue without changing the name; in order
to facilitate this, we select the name in the Text control so that the
first keystroke replaces it completely.
Not suggesting any name in this case is probably not as good as the
user might start to type in the same name as would have been suggested
just to find out that the branch already exists.

Bug: 334256
Change-Id: If60302728635e4a7ab2b824be0c24e003f2fff2c
Signed-off-by: Benjamin Muskalla <bmuskalla@eclipsesource.com>
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>