Update patch set 1
Patch Set 1:

A problem with @SupressWarning is that it is too coarse.This would be the first deprecated warning to suppress and I'm not convinced it is a good idea. We are programming for the 3.5 API and you will get different warnings depending on the Eclipse version you are running. This differs from other warnings, that do not depend on our Eclipse version. I.e. we know better what is hidden.

The SupressWarning annotation should perhaps refer to the use of addDerived, because anyone looking at the code with Eclipse 3.5 will not understand the problem. Replacement for what?

You can hide warnings in the problem filter.

Patch-set: 1