Update patch set 1
Patch Set 1:

I think extensions shouldn't be able to break the commit dialog, hence if they throw exceptions we should fallback to the un-extended commit dialog to ensure that this important dialog also works if buggy extensions are hooked on.

Yeah we should also provide hooks, but I feel this should somehow involve jgit, since we may also want to have the ability to install hooks on jgit level if jgit is used in a different context than EGit.

Patch-set: 1