Update patch set 1
Patch Set 1: (1 inline comment)

What is my intention of the usage of this extension point:

In an own plugin org.eclipse.egit.mylyn I can implement the ICommitMessageProvider:

public class MylynMessageProvider implements ICommitMessageProvider{

   public String getMessage(){
      ITask task = TasksUi.getTaskActivityManager().getActiveTask();
      
      String comment = 
           ContextChangeSet.getComment(...);
      return comment;
   }
}

This class is the messageprovider I needed to configure in plugin.xml. So I am able to add the mylyn's generated commit message for the currently activated task.

I have another idea to add more functions for a really task-focused working with eclipse, git and mylyn.
If you are interested we can discuss this on the dev list.

Patch-set: 1