Update patch set 2
Patch Set 2: Do not submit

I'll push a better version of this one that takes care of ambiguous ref selections.

The reference to a failed build was a local address. Egit builds currently fail because I broke the JGit API, but only the JGit part was merged and not the one needed to fix EGit http://egit.eclipse.org/r/#change,2178

Patch-set: 2