Simplify Fetch and Push second part: Add a simplified Push wizard
A new pair of "Push to upstream"/"Fetch from upstream" menu actions is
added to the team menu as well as the Repositories View. When clicked,
the relevant remote is detected (either upstream configuration of the
currently checked out branch, or "origin"), and the push/fetch is
tried for this configuration if possible.
If the currently checked out local branch is configured to use "." as
upstream (i.e. the local Repository, because the branch is based on
another local branch), the actions are disabled.
In any case, after the fetch/push, the user can click "Configure..."
on the push/fetch result dialog in order to inspect or change the
configuration.
Alternatively, the user can use the new "Configure Fetch from
Upstream..." and "Configure Push to Upstream..." menu items in the new
"Remote" sub-menu available in the Team context menu to do the
upstream configuration.
Also, the "Fetch from..." and "Push..." actions (with the advanced
wizard) were moved to that new sub-menu to reduce the size of the Team
menu.
When configuring the Push, there is a small wizard that allows to
easily select the correct branches etc. For the Fetch case, there is
currently only a rather technical dialog, which should be replaced by
some wizard asking questions like "Do you use Gerrit to push to" in
order to let the user make the right decisions.

Bug: 312586
Bug: 313862
Change-Id: I03c9010717117eaa4f6f00ceee20a5037c024e9e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>