[historyView] Missing adapter call in command handler
The History View Command handler must properly adapt input objects
that are not IResources, otherwise a ClassCastException will be
thrown. This was reported by a colleague to happen when opening a Java
file in the history view and then trying to "Reset" onto another
commit.
Even though I couldn't reproduce this exact scenario, the code must be
fixed anyway.

Change-Id: Ia4722eb182909108f30828901a5fdd9f4390599c
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>