Check HOME environment variable on Windows on first usage of EGit
The home directory is calculated by concatenating HOMEDRIVE and
HOMEPATH if HOME is not defined. This can result in follow-up problems
e.g. the unwanted usage of a network share as home directory.
The user is shown an information dialog the first time EGit is used if
the HOME variable is undefined.

Change-Id: Ieb3d242de135e692fb42dc8054f4a29ddffbfc4b
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>