CommitDialog: improve sorting of file list
This initially sorts the file list by status, project name, and
project relative path.
When clicking on either the status or file column header, the list is
re-sorted according to the clicked column ascending (in case of the
file column by project and project relative path).
The second click on the same column header re-sorts according to the
clicked column descending. The third click on the same column reverts
the original sorting (by status, project and path).

Change-Id: Ie027d22b17b0b8afcff4098851ea39d56d79137c
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>