Update patch set 3
Patch Set 3:

As for remark 1., I see this as a valid requirement, but we can add this later, I think. Then we would probably want all the other nice features, like search, toggle between "Show All" and "Normal" history and so on.
As for remark 2., I have rebased this on the other change which fixes this.

Patch-set: 3