Update patch set 3
Patch Set 3:

Dariusz, you're right, the patch doesn't solve all problems, but compare editor for non-workspace files is a bit different story. The bug report was that we can't use Sync View at all (it's always empty). And with patch we can compare files those are inside of project (I think most common usecase). I will go on working to fix compare for non-workspace files, this is ofcorse also important, but can it be a different patch? We want to start using EGit in production and this fix is critical.

Patch-set: 3