Update patch set 5
Patch Set 5:

What do you think about the idea es below?

We could check for an optional file in the /.settings/ directory (where eclipse related configuration files are) and read the hostname/URL from that file.

The file could be for example: /.settings/org.eclipse.mylyn.tasks.core.TaskRepository

The advantage of this solution is that this file will be also copied during cloning.

Patch-set: 5