Streamline Import Wizards first part: get rid of manual import
Since the Git Sharing Wizard now is able to handle multiple projects
that are to be shared with the same repository nicely (i.e. you don't
have to share these projects one by one as before), we can drop the
"share manually" option in the Git Import wizards. This makes these
wizards much less complex and avoids the need to complete them in the
UI thread.

This is the first step for simplification which removes the page that
is used to manually share newly created projects.

Change-Id: I0c68173eb83f2250e0d0b4526ec184dbe234a927
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>