Move "Compute Change-Id for Gerrit" to repo config
The idea of this change is that we can simplify the commit dialog by
removing the Change-Id check box from it. Instead, it can be configured
in the Git property page of a project. The setting is stored in the
.git/config file of the corresponding repo as follows:

[gerrit]
	addChangeId = true

In the commit dialog, we then add a Change-Id automatically if it's
configured for the repo to which we are committing. If the Change-Id is
not wanted for a specific commit, it can be removed by deleting the line
from the message.

Change-Id: Ie1b49677cf2cec30a0a41ac15608db9dd87b9dc6