Simplify Fetch and Push: configure "Pull" when creating branch
This allows the user to configure "Pull" during branch creation. The
default configuration as in autosetupmerge and autosetuprebase is
taken into account and the user can override this.

The Branch Creation wizard page has become a bit more complex due to
this, but the user should be able to work just fine with the defaults.

There is now an information message telling the user that they
normally want to base their branches on remote tracking branches.

The remote tracking branches now appear first in the list of available
branches to also support users in this.

Note that the underlying JGit API operation does not yet support the
autosetup* flags yet.

The PullResultDialog now is able to show a Rebase result, too.

Change-Id: If3cb758b8c04039219cbb090c2385ca5315ad2df
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>