Update patch set 2
Patch Set 2:

I like Matthias' idea. We could still add a context menu in the Repositories View to quickly open this preference page (without having to click Window > Preferenes > Team > Git > Configuration > Select Repository), but in general I think that this would be the right place to add this UI.

It shouldn't be a big deal to add this third tab to the Preferences page (something like "Repository setting" next to "User setting" and "System setting").

Unfortnuately, we don't have any semantics on these pages (yet), so the next step may be to improve the dialog shown after clicking the "New Entry..." button with some help regarding keys and values to assist users with the task of setting configuration values.

Patch-set: 2