Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I wonder if we shouldn't put things where they belong: if this is a repository configuration, we should show it there (and let it be changed there, too).
Unfortunately, the UI support for the repository configuration is currently very limited (it's shown as Property page when a Repository is selected in the Repositories View), but we will need more of this kind of configuration soon (autosetupmerge and autosetuprebase come to my mind, for example), so  I think we should look into this direction (e.g. some sort of "Repository configuration" dialog).
I don't think Eclipse users would ever understand that they have to open a Project property page to find a property that is not stored in the project meta data.
Furthermore, going the configuration approach, we could also tell users to add this flag to the user-specific configuration so that they don't have to bother with configuring each repository one by one.

Patch-set: 2