[repoView] avoid repeated logging of Exception during startup
The RepositoriesView looks up the RepositoryCache with all the
directories it "knows". When an Exception occurs, it is logged, but
the corrupt Repository is not removed from the view. When the view is
refreshed or re-initialized, the Exception is logged again.

The Exception described in the bug occurs if the config file does not
contain an entry

core.repositoryformatversion = 0

which is checked in the constructor of FileRepository. The check was
introduced to ensure interoperability with the command line Git.

With this fix, the Exception is only logged once and the offending
Repository is removed from the Repositories View.

Bug: 318875
Change-Id: I234c94f4f1a053b441e86f9e1aeaf55cf3ca4587
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>