Update patch set 1
Patch Set 1:

For me, doing it this way also doesn't feel quite right. It's not a preference for Git or Eclipse, but belongs with the repository. I'd prefer it to be just available from the Git Repositories view as an entry in the context menu which opens a property page for configuring the selected repository.

Patch-set: 1