Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Very nice already. I would just want to have a label for the "Repository" combo.

In addition, I think we need to be able to toggle between "effective" and "repository-specific" configuration. At the moment, you see the effective config, but you shouldn't be able to edit this, because it would be unclear which file fould be affected by a change (global, user or repo-specific).
I imagine a checkbox ("Show effective configuration", default off, when checked, the edit buttons should be disabled).

I must say that I'm not quite sure about whether having this in the Preferences area is the right thing to do, but I wouldn't mind having it just like you suggeted it.

I agree that it's not "Configuration" in the classical Eclipse meaning, but then, the Repo config is tightly integrated with the global and user configs, so I think it dramatically improves usability.
 
On the other hand, creating a dialog showing the same information shouldn't be too much effort (at the cost of loosing the integration aspect). This could then be added to the Repositories View as menu action.

In any case, I think we can put this discussion aside for the moment. Refactoring this out of the Preferences afterwards shouldn't be too much effort.

Patch-set: 1