Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

When clicking "Commit" from the repository node in the Repositories View "modified, not staged" files under this repo's working tree aren't checked by default in the file list of the commit dialog. This is not consistent with the behavior of the team menu in the package explorer. I think they should behave in the same way as the input for the operation when calling it from the repository node is the whole repository.

Patch-set: 3