[historyView] Performance second part
This limits the number of commits to show in the history page to ten
thousand. This is a UI Preference available on the Git Preference
page. The user can set this to zero in order to see all commits.

The second thing is to hide the "tag sequence" by default. This is
very time and memory-consuming on big repositories. This is changeable
by another Git Preference.

In combination, this makes browsing the Linux 2.6 repository work
without too much waiting and memory overflow.

Bug: 337081
Change-Id: Ie2823c3c24c2ee9a1a9baf1a6c314901bf8b4a12
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>