Update patch set 1
Patch Set 1:

Right now, this change causes the password to be stored in plaintext inside of .git/config.  That is far from optimal, as this file is often world-readable.

We need to either extract the password and store it in ~/.netrc (and honor reading from ~/.netrc) to be compatible with C Git, or we need to start using the encrypted password storage within Eclipse like the CVS team provider uses.  Or both.

Patch-set: 1