[RFC] [historyView] Show full Ref name when hovering over label
Requires JGit change

http://egit.eclipse.org/r/#change,2556

Displays the full name of a Ref when hovering over the label in the
HistoryView. This is helpful when the Ref name gets too long and is
shortened during rendering of the commit.

I have tagged this as RFC as the hover sometimes seems to conflict
with the tooltip shown if the width of the message column is too
narrow to show the full message. Perhaps someone has an idea how this
can be avoided.

Change-Id: Ia29e2eafb9413d6adb4e27b398d71f5053f2beb2
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>