[sync] Add tests for GitChangeSetSorter
This change adds tests for GitChangeSetSorter. It also fixes minor
issues in this class that were found during writing tests.

GitChangeSetSorterTest is fully independent from Eclipse RCP therefore
it can be launched as pure JUnit4 test case.

MockRevCommit class is added to be able to set commit time without
setting up a full git repository, therefore
GitChangeSetSorter.shouldOrderCommitsByCommitDate() will fail with
ClassNotFoundException when this tests is launched as a plug-in test.

Change-Id: I6b5ca33c042c935bd299a956858528d0bea2989a
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>