Create temporary decorations for project nodes
The label decoration of project nodes contains important information,
i.e. the repository name as well as the current branch. This change
introduces a temporary decoration for project nodes that only contains
this textual (label) information and is shown immediately. The final
decoration is shown as soon as a full recursive walk through the whole
project has been performed.

Change-Id: I376b9284b35bfc5960102182753ed90498da0829
Signed-off-by: Philipp Thun <philipp.thun@sap.com>