Show commit corresponding to selection in commit graph table
Let the commit graph selection follow the Ref selected in the
Repositories View so that the corresponding commit is visible in the
History View. If a resource is selected HEAD will be made visible.
This is done if History View is configured to follow selection and
filter is set to show all changes in the repository.

Change-Id: I81f8edac2c3c14bb428524b7d0cf2e7e8d108905
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>