Update patch set 1
Patch Set 1:

Just take the commit you made, amend it and edit the commit message (and the files, if you want). Then commit and push again (make sure, that the Change-Id in the commit message is the same like the one from this change, that should automatically happen if you amend the right commit). Then changed commit should be added as Patchset 2.

Patch-set: 1