Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(6 inline comments)

I like the idea to open up EGit by offering extension points. The problem is that we have not yet really discussed about a concept of a public API. At the moment all packages are marked as internal.

I still do not quite understand the use-case to manipulate the set of files to be committed. This would be very intransparent to the user.

Patch-set: 2