Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This patch works great for workspace files, but for non-workspace files it opens an editor with information: "Could not open the editor: The file does not exist.".

IMHO the "Open" action should support opening non-workspace files or at least should be disabled for those kind of files.

Patch-set: 1