Implement project rename and move for some cases
We can now move and renme projects within a repository for a number
of cases. Moving a project to a subdirectory does not work and since
Eclipse actauly destroys the project, we simply forbid it for now
(Platform bug 339814).

This also adds tests for move and delete of folders and files. We are
quite inconsistent so the current behavior may not be the final behavior.

We cannot move projects with repositories inside them in this version and
so we forbid it. Disconnect and share manually to accomplish this.

Bug: 334112
Change-Id: I634d5b44bc7fc6a7453b5d718733c9f8f3542840
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>