[RFC] Allow variables for default clone destination
Using the user's home directory as the default clone destination of
repositories is a confusing choice in Eclipse. The CVS and Subclipse
plugins use the workspace location, and the new project wizard also
defaults to the workspace.

This change allows to use variables to set the default clone
destination in the preferences. It also changes the default setting
to ${workspace_loc}.

Change-Id: I5f1cb2c2c38d308a0a67c60b9546b6a5673bca08