[RFC] Enable pull for multiple repositories
When having multiple projects (each corresponding to a different
repository) and needing to regularly keep up to date with changes, EGit
provides no easy way to deal with this. The only way at the moment is to
select a single project and pull, and repeat that for every project.
When the number of projects is two, this is okay, but for more than a
couple of projects it's simply infeasible.

This change enables pull for multiple repositories. It's more a proof of
concept and has multiple problems, but it's there to get the discussion
started on how to implement this.

Bug: 340308
Change-Id: If8a51bb9ad0c78dfeb73ec1f4ad0b4602c6feb1f