Enable pull for multiple repositories
When having multiple projects (each corresponding to a different
repository) and needing to regularly keep up to date with changes, EGit
provides no easy way to deal with this. The only way at the moment is to
select a single project and pull, and repeat that for every project.
When the number of projects is two, this is okay, but for more than a
couple of projects it's simply infeasible.

This change enables pull for multiple repositories. The UI still has to
be improved in a subsequent change, as it currently shows one result
dialog per repository. It has to be changed so that only one result
dialog is shown, see bug 340780 for that.

Bug: 340308
Bug: 340780
Change-Id: If8a51bb9ad0c78dfeb73ec1f4ad0b4602c6feb1f
Signed-off-by: Robin Stocker <robin@nibor.org>