Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Hmm... I'm afraid I disagree, at least with the suggested implementation. The RepositorySelectionPage is used all over the place, and when using it e.g. in the clone wizard, the URI field is clearly the one that should get focus (the remote selection combo wouldn't even be visible there and you'd get a NullPointer with your code).

I'm not quite sure how this should be done. Perhaps you need to check if the remote field is active (a new method isRemoteSelected() similar to isUriSelected()) or something.

BTW, you shouldn't have to go through the push wizard every time you push to Gerrit. Just add the Gerrit URI as push URI and the correspondig push spec to your "origin" spec and hit "Push to Upstream"... from the team menu or repo view context menu on the repo node. As long as your local branch is based on origin/master, all should be done for you automatically.

Patch-set: 1