Fetch and Push wizards: allow to finish on first page
If the user selects a remote (and not a URI) for push and fetch,
respectively, the corresponding wizards should allow to hit "Finish"
on the first page if the remote is configured properly (i.e. has at
least one URI and a ref spec (for fetch).

The bugs mentioned below are more concerned about the complexity of
the RefSpec page, so if the remote is not configured correctly, this
does not solve the issue, but for the 90 per cent case, the user will
be able to simply hit "Finish" without having to go through the other
pages. This also dramatically improves performance, as the
long-running ListRemoteOperation is not needed anymore to fill the
RefSpec combo boxes on the RefSpec page.

Bug 333410 should be solved by this.

Bug: 312586 (only partial solution)
Bug: 313862 (only partial solution)
Bug: 333410

Change-Id: I85b03db3931192d0b6764cc056fd5e7a5da44450
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>