[repoView] Add scoped keybindings for copy and paste
Show key bindings for copy/paste repository path
context menu action that appear in Git Repositories view.

Created a context for that view and added key bindings for those two
actions to the plugin.xml file.

This now allows those key bindings to appear and be configured in
the Keys preference page as well as have a context available for
"In Git Repositories View".

Bug: 331484
Change-Id: I508a0ed4581a9619e13cb6698ce123058ae0f9ec
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>