Update patch set 3
Patch Set 3: Looks good to me, approved; Build Successful; IP review completed

Nice feature. The only thing is missing is that it must be possible to activate the pop-up again. This should be possible using a Git Preference page. Since we need to reorganize these pages anyway, we can do this separately. I opened bug 340630 for this reorg stuff and mentioned this change there, too.

Patch-set: 3