Update patch set 3
Patch Set 3:

Unfortunately this is not true (see screenshot I sent via separate email), I suspect that there is a problem with the menu activation rules in repositories view if the repository is empty.

I thought we could show the information that the repo is empty on the label decorator of the repository (in repositories view) / 
project level (package explorer).

Patch-set: 3