Implement a "Switch to..." action
This adds a dynamic "Switch to..." sub-menu to the team menu that
allows to switch directly to any local branch or (with an "Other...")
button allows to open the current "Branch..." dialog to select a tag
or remote tracking branch.

The number of dynamically created items is limited to 20, but the user
can still open the "Other..." sub-menu to switch to a local branch
that is not in the list or to a tag etc.

Another action "New Branch..." opens the CreateBranch wizard so that
the user can create and check out a new branch directly.

Bug: 336931
Change-Id: I8d03c88b7a578f86eb63b61fda19584f78e38f22
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>