Change ref parameter of CloneOperation to String
The parameter is used for specifying what branch should be checked out
after fetch. Before this change, it was of type Ref. But a Ref is not
really right here, as a Ref should point to a valid object id, and the
ids aren't generally known before a clone. This lead users of
CloneOperation to passing a Ref with only a name set, and no object id
(using a anonymous Ref subclass or passing null as object id).

Before the parameter was a Ref, it already was a String, but one could
only specify a short branch name (e.g. "master"), not a full ref name
("refs/heads/master"). With this change, it's still the full ref name,
just as a String.

Change-Id: I420059e7eda4b3bbbf335557217bc89f11c27abd
Signed-off-by: Robin Stocker <robin@nibor.org>