Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

A few comments - apologies for lack of inline but iPhone doesn't permit me :-)

You're limiting selection to one project. Why? If they are all the same repository then there's no harm. If there are multiple repos then we might disable, or could show common names. For example, I might naively think I have to select all EGit projects to switch to another branch, or in the future, you might want to switch all EGit and JGit projects to the 0.11 branch. Although the branch would be different in each case they have the same name. 

The entry set could give branches in arbitrary order if I read this correctly. If so, can we sort?

The comment in the single arg constrictor has "default constructor" which seems to be a cut and paste from the single arg version. 

The other menu should probably be disabled in the case of multiple projects being selected with individual repositories.

Patch-set: 1