Update patch set 4
Patch Set 4:

Regarding documentation: out documenation is not checked into the respository, but maintained in the wiki and copied from there when a new version is released. Normally, we do an update of the wiki shortly before the next version is released thus avoiding too much discrepancies between the current release and the wiki.
So yes, we should update the documentation, but not just yet.

Patch-set: 4