Refactoring Branch operations second part: specific dialogs
This adds some dialogs to be used for checkout and deletion of
branches which are used in the Team->Switch To...->New branch... and
Team->Switch To...->Other... Menu actions.

New Branch... will first call a dialog to select the base (branch,
tag, Reference) and then start the usual wizard with the selected base
as parameter.

Other... will open a dialog to select a branch, tag, or Reference for
checking out. Upon "Ok", the Ref is checked out.

With this, users can use specific, simplified UIs for the very common
operations of branch creation and checkout instead of having to use
the generic BranchSelectionDialog which offers too many actions.

Change-Id: I39cb1a8f97c02e419f37a864951208cee70ef62e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>