[RFC] Refactoring Branch operations third part: cleanup
Add a "Rename Branch" dialog and remove the BranchSelectionDialog. Use
an "Advanced" sub-menu in the Team menu to add actions for "Delete
Branch and "Rename Branch" and remove the old "Branch..." action from
the Team menu.

Move the "Tag..." action also under the "Advanced" sub-menu.

Change the "Branch..." action in the action set to use the new
"Checkout Branch" dialog.

In a subsequent change, we could also move the "Assume Unchanged/No
Assume Unchanged" actions to the "Advanced" sub-menu.

Change-Id: I7a50c4c0f6c42e9ed4e9d1d18d6587b0f5525d3c
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>