Update patch set 3
Patch Set 3:

I will work on it, but can't give any estimates as to when I'll be able to submit it.

In the meantime, it would be nice if this simple implementation could get in. IMO it's better than no support for pull multiple, and it doesn't negatively affect the interaction for pulling one.

Patch-set: 3