Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(2 inline comments)

Calling the pull operation several times makes the user experience a bit ugly, for example when it comes to errors (multiple error pop-ups etc.)... Also the result dialog currently would be shown several times... Not sure how problematic this is in reality. In the long run, we might need to improve PullOperation to accept multiple repos in its constructor and make sure error handling.

In any case, I'd like to get a second opinion on this aspect.

Patch-set: 1