Team project set support
This is based on the previous work by Mykola Nikishov in change
I800dd9f23953ad903f20289c442707e2dde08f75 ("Import a project set file").
This change includes export support, unit tests, and can handle the
situation where different branches of the same repository should be
imported.

A project reference string in the project set file looks like this:

  1.0|git://egit.eclipse.org/egit.git|master|org.eclipse.egit

Based on an array of these reference strings, GitProjectSetCapability
does the following when importing:

- Clone a repository with CloneOperation for each combination of
  repository URL and branch name.
- Import a project from each directory that is specified in the last
  part of the string ("." if the repository itself is the .project).
- Connect the projects using ConnectProviderOperation.

Due to Bug 244313, import and export of project set files requires the
org.eclipse.core.runtime.compatibility plug-in. The dependency is in a
separate optional feature, as proposed on the old change. It's required
for Eclipse < 3.7.

Bug: 296082
Change-Id: I1500cc376961117b39b8760e6de675a4885e5f72
Also-by: Mykola Nikishov <mn@mn.com.ua>
Signed-off-by: Robin Stocker <robin@nibor.org>