Update patch set 1
Patch Set 1: Build Failed

Not sure if this is the right thing to do. Depending on the use case, you might rather expand the local branches node...
I suggest to either offer a method on the AbstractBranchSelectionDialog to expand some or all nodes (similar to setRootsToShow, maybe the new method could be called setRootsToExpand) and call this from the concrete subclass (in this case the RebaseTargetSelectionDialog).

Patch-set: 1