Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I was holding off on the plugin.xml registration of the editor input factory until the commit editor was also registered in the plugin.xml as an editor.  I think it makes sense to have the plugin.xml changes for the editor and input factory being the same commit.

Patch-set: 2