[RFC] Add a 'Staging' view
Git's staging area is one of its most powerful features. This view
provides users with the ability to drag-and-drop files between
the working directory and index.  Such an operation would be
synonymous to adding the modified files to the index and tracking
the untracked files. Other operations like committing is also
possible from the view's context menu.

Change-Id: I7d78aff4db32d864ac0410a178bad149c45b6440
CQ: 5115
Bug: 313263
Also-by: Remy Suen <remysuen@ca.ibm.com>
Also-by: Matthias Sohn <matthias.sohn@sap.com>
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>