[DRAFT] task activation integration support
This change enables the automatic creation or checkout of branches
according to the activated mylyn task. If there's no context in the
task, a dialog appears, where the repositories can be selected. If
there's a context, the resources of that context are browsed and the
contained Git repositories are selected for branch creation/checkout.

Bug: 309578
Change-Id: I03b927ab8701caa13ec62e536ebdb351d11439e6
Signed-off-by: Manuel Doninger <manuel.doninger@googlemail.com>