Allow committing non workspace changes
The commit process is now based on files instead of resources. It is
possible to commit changes on files that do not belong to the
workspace. The commit dialog commits the changes of one single
repository.
Displaying diffs in the commit dialog for non workspace files needs to
be implemented in a follow-up commit.

Bug: 329061
Bug: 302349 
Change-Id: Ib5f0bb2a6d86000683415b100702a0b322efd032
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>