Update patch set 2
Patch Set 2:

We should check for that and only then disable it (or enable only do the push for those repo where it's possible).

It's possible when all repos have a remote configured and are not compared to a local branch (IMO, comparing to remote and pulling/pushing will be the primary thing people do with the synchronize view once it's possible.)

We should aim to make the synchronize view at least as good as the CVS one. And the CVS synchronize view *can* commit to multiple modules/repos at once. That's one of the reasons why it's so useful.

Patch-set: 2