Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

I would also prefer to have the same UI in both the Team -> Synchronize... dialog and the Synchronize wizard. In the Team -> Synchronize... dialog it is clearer whether I select a branch or a tag.
But in the synchronize wizard I have to deal with more than one repositories. I do not know yet how to combine the UIs best.

We could in a first step leave both UI as they are. In the current wizard there are some problems:
- The finish button is enabled even if either source branch or destination branch is not selected
- Source branch selection is not disabled if "Include Local" is selected

Btw.: I think that we have the wrong order of "source" and "destination" in both the UIs. Source should be the remote one, right?

I'll add Kevin as reviewer. He may have some good UI layout ideas

Patch-set: 3