Update patch set 3
Patch Set 3:

How about having the left column of the Synchronize wizard
table be project driven instead of repository driven.

So it would now be a table instead of a tree with projects
shown as:

org.eclipse.egit.ui [egit master]
org.eclipse.egit.core [egit master]

intead of:
/home/ksawicki/repositories/egit/.git
---org.eclipse.egit.ui
---org.eclipse.egit.core

And set the default branch in the table to be origin/master
similar to the Synchronize... option.

Patch-set: 3