Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(1 inline comment)

Aside from the minor inline remark, I'd like to challenge the use case: what is the added value for the user of the view to see the number of tags and remotes?

They already have the "hasChildren" indicator next to the respective node, so they can already see whether there are tags or remotes and we should be careful not to clutter the tree with pointless information.

Patch-set: 1