Update patch set 2
Patch Set 2:

I doubt this is really useful, to my experience both the number of tags and remotes isn't changing often so what's the point in seeing these numbers ? What's changing much more often is the number of local branches but I also don't see why the number would be important. If certain platforms don't show if the node has content or not we should rather file a bug for SWT to fix this.

Patch-set: 2