Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

(34 inline comments)

Why did you remove all "@see" javadocs? It it is necessary for some reasons this should be done in separate change or at least mentioned in commit message.

Also I think that 'Expand All' and 'Collapse All' buttons should be disabled (if it is possible) because there is thing to be collapsed/expanded in search result.

Apart from that, this is a very good idea to have search commit option! I like it ;)

Patch-set: 3