Update patch set 1
Patch Set 1:

Well, for the unstaging feature, I'll likely submit the change set to Gerrit separately and wait until that gets merged before I integrate that code into this code base (adding the extra DND hook).

As to the 'Synchronize' view, as I mentioned on bug 309582, I do not feel that the 'Synchronize' view is the right place to expose the staging area but if you all think so that's fine by me. Sadly, it seems only Stefan has been chiming in on the discussions. :/

At any rate, I'm not the one that's using Git on a regular basis, I use CVS at work. :) I just want to compare files easily in a view without having to bring up the commit dialog and then double-clicking on those entries individually. ;) I don't really mind where this feature comes from as long as it's accessible from a view.

As an aside, we are really late in the endgame already here in my opinion. Since this change is pretty big, I don't see this feature going into 0.8 so there should be plenty of time to bounce ideas around.

Patch-set: 1