Update patch set 1
Patch Set 1:

IMHO this change with unstaging files should be included in this patch set ...

Another thing that I'm not really happy of is that we are getting next view (we have already 'Git Repositories', in nearest feature we would have 'Syncrhonize view'). It seams that if user want to use Git integration in Eclipse he should have opened two or tree additional tabs. This could be quite anoing.

I'm wondering if we can merge both view (I mean this one and 'Synchronization') ...

Patch-set: 1