Update patch set 1
Patch Set 1:

Agreed. Right now the view pretty much exposes nothing more than what you already see [in EGit]. Besides the fact that removed files show up, it does not provide any additional actions at all since everything you see in the context menu is the same as if you right-clicked on a file or a project in another view. :)

A separate change set should be created for allowing the resetting of individual files and then this view will get it for free in the context menu...though some extra code will be needed for the DND case. ;)

Patch-set: 1