Update patch set 1
Patch Set 1: No score

I play a little with this feature and first thing that I've notice was that, I can't unstage file after it is staged. I've try to drag it into 'Modified' section but it won't even move.

IMHO there should be also buttons for staging and unstaging so that we user can use 'Staging' with drag and drop action.

Patch-set: 1