Update patch set 1
Patch Set 1: No score

Dragging projects/folders to the other nodes won't work. I was trying to implement this at one point but the logic was getting kind of crazy...or I was approaching the problem incorrectly. Anyway, that's definitely something that needs to be implemented at some point.

Patch-set: 1