Update patch set 1
Patch Set 1:

In deed it looks that in [EJ]Git we can't done reset for a single file, but only for a project. This should be changed IMHO.
I didn't look on EGit 'reset to' implementation but maybe we can extract some part form it and implement 'reset to' for a single file in 'Stage' view.

Patch-set: 1