Update patch set 3
Patch Set 3: No score

Honestly, I'm not sure a CQ is necessary. Figure 1 makes a distinction between "Submitting Committer" and "Committer on same Project". To me Remy is a "Submitting Committer" because he already *is* an Eclipse committer. He is not a committer on EGit/JGit but he is an Eclipse committer. 

I'll get a confirmation on my interpretation and get back to you.

Patch-set: 3