Update patch set 3
Patch Set 3:

Gunnar, I got the image from Team's plug-in or the CVS plug-in.

Stefan, you're right, we don't have an action yet for that. That seems like another action we'd have to implement. Though at the moment, it's not possible to commit only staged changes...so I guess we need an action for that too. ;) I thought the modified/staged entry would have modifiers but maybe my memory is off. I'll have to try it later. The 'Compare With...' is just contributed from the EGit code. Everything from the context menu is standard stuff from Eclipse or EGit. We'd have to correct them individually.

Robin, it should be only listing "dirty" files. Could you rephrase your comment? The view is pretty much the same as a view of your workspace, except that it's filtered for dirty and staged files. Showing files outside the Eclipse workspace I guess would be doable. That's also a problem the synchronization feature would need to consider now that I think about it.

Patch-set: 3