Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Actually JVM uses StringBuffer (on bytecode level) for all append operations on strings.

Lets suppose that in feature somebody will implement better append algorithm than it is actually used in StringBuffer and this new implementation would replace StringBuffer in complied java classes. If we are bind to StringBuffer implementation in our code we should replace it, but if we would be relaying on JVM string-appending-implementation we'll get performance boost after code recompilation.

Therefore I'm not sure that should we tray to be more intelligent then java compiler and JVM developers ...

Patch-set: 1