Update patch set 2
Patch Set 2:

Dariuz, the culprit was the mapping as it did not return exact traversals (only projects) for the resources included in this model container.

Regarding the resources outside of the workspace: I'm not sure if we should handle this case here. Maybe the commit dialog should warn the user that there may be resources outside of the workspace.

Patch-set: 2