Update patch set 6
Patch Set 6:

I think this is the preferred layout. So let's start with this and only add options when this seems to be necessary after the new view has been used in the wild for a while. 

In a later step we should add another gitx-ish diff view enabling staging and unstaging of individual hunks (equivalent to git add -i). This could probably reuse the diff view implementation of Kevin's new commit viewer which has been integrated into the History View context menu yesterday. 

The width of the commit message pane when resizing the view should be limited to the width signaled by the vertical line showing where the automatic wrapper will start wrapping the lines and the file path panes should grab all remaining horizontal space. This way we gain maximum width for the typically lengthy file paths.

Patch-set: 6